requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
configparser==5.0.0
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
multipledispatch==0.6.0
numpy==1.17.4
opencv-python==4.4.0.44
pillow==8.1.1
pyrr==0.10.3
pyyaml==5.4.1
redis==3.5.3
scipy==1.5.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
torch==1.6.0
torchvision==0.7.0
